+++ src/cosmos_bio_cns/__init__.py
@@ -8,6 +8,7 @@ from cosmos_bio_cns.fusion import BioFusionEngine
 from cosmos_bio_cns.models import BioFeature, BioObservation, CNSState, ConsentScope, FusionFrame, HeartbeatRecord
 from cosmos_bio_cns.persistence import SQLiteEventStore
 from cosmos_bio_cns.runtime import BioCNSRuntime
+from cosmos_bio_cns.schema import load_schema
 
 __all__ = [
     "BioAdapter",
@@ -24,6 +25,7 @@ __all__ = [
     "OrganStatus",
     "RunningBaseline",
     "SQLiteEventStore",
+    "load_schema",
 ]
 
 __version__ = "0.1.0"